# packages/contracts/openapi/influaudit.yaml
components:
  schemas:
...
    auth.AuthResponse:
      properties:
        access_token:
          type: string
        expires_in:
          type: integer
        refresh_token:
          type: string
        token_type:
          type: string
        user:
          $ref: "#/components/schemas/auth.UserResponse"
      required:
      - access_token
      - expires_in
      - refresh_token
      - token_type
      - user
      type: object
    auth.LoginRequest:
      properties:
        email:
          type: string
        password:
          type: string
      required:
      - email
      - password
      type: object
...
    auth.RefreshRequest:
      properties:
        refresh_token:
          type: string
      required:
      - refresh_token
      type: object
    auth.RegisterRequest:
      properties:
        email:
          type: string
        full_name:
          type: string
        password:
          type: string
      required:
      - email
      - full_name
      - password
      type: object
    auth.UserResponse:
      properties:
        created_at:
          format: date-time
          type: string
        email:
          type: string
        email_verified:
          type: boolean
        full_name:
          type: string
        id:
          format: uuid
          type: string
        role:
          type: string
      required:
      - created_at
      - email
      - email_verified
      - id
      - role
      type: object
...
    billing.SubscribeRequest:
      properties:
        plan_code:
          type: string
      required:
      - plan_code
      type: object
    billing.SubscriptionResponse:
      properties:
        cancel_at_period_end:
          type: boolean
        current_period_end:
          format: date-time
          type: string
        current_period_start:
          format: date-time
          type: string
        id:
          type: string
        plan_code:
          type: string
        plan_id:
          type: string
        status:
          type: string
      required:
      - cancel_at_period_end
      - current_period_end
      - current_period_start
      - id
      - plan_code
      - plan_id
      - status
      type: object
    billing.WebhookRequest:
      type: object
...
    influencer.AddHandleRequest:
      properties:
        follower_count:
          nullable: true
          type: integer
        handle:
          type: string
        platform:
          type: string
        platform_user_id:
          nullable: true
          type: string
        verified:
          type: boolean
      required:
      - follower_count
      - handle
      - platform
      - platform_user_id
      - verified
      type: object
...
    influencer.HandleResponse:
      properties:
        created_at:
          format: date-time
          type: string
        follower_count:
          nullable: true
          type: integer
        handle:
          type: string
        id:
          type: string
        influencer_id:
          type: string
        last_seen_at:
          format: date-time
          nullable: true
          type: string
        platform:
          type: string
        platform_user_id:
          nullable: true
          type: string
        updated_at:
          format: date-time
          type: string
        verified:
          type: boolean
      required:
      - created_at
      - handle
      - id
      - influencer_id
      - platform
      - updated_at
      - verified
      type: object
...
    report.PublicBadge:
      properties:
        authenticity:
          nullable: true
          type: number
        benchmark_label:
          type: string
        generated_at:
          type: string
        handle:
          type: string
        niche:
          type: string
        overall:
          type: number
        pdf_url:
          type: string
        tier:
          type: string
        verification_tier:
          type: string
      required:
      - overall
      type: object
...
    waitlist.CaptureRequest:
      properties:
        email:
          type: string
        influencer_id:
          type: string
        props:
          format: byte
          type: string
        source:
          type: string
      required:
      - email
      - source
      type: object
    whitelabel.UpdateWhitelabelRequest:
      properties:
        brand_name:
          type: string
        custom_domain:
          type: string
        logo_url:
          type: string
        primary_color:
          type: string
      required:
      - brand_name
      - custom_domain
      - logo_url
      - primary_color
      type: object
    whitelabel.WhitelabelResponse:
      properties:
        brand_name:
          type: string
        custom_domain:
          type: string
        logo_url:
          type: string
        primary_color:
          type: string
        updated_at:
          format: date-time
          type: string
      required:
      - brand_name
      - custom_domain
      - logo_url
      - primary_color
      - updated_at
      type: object
info:
  title: InfluAudit API
  version: 0.1.0
openapi: 3.0.3
paths:
...
  /auth/login:
    post:
      operationId: Login
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/auth.LoginRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/auth.AuthResponse"
          description: OK
...
  /auth/refresh:
    post:
      operationId: Refresh
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/auth.RefreshRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/auth.AuthResponse"
          description: OK
  /auth/register:
    post:
      operationId: Register
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/auth.RegisterRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/auth.AuthResponse"
          description: OK
...
  /billing/subscribe:
    post:
      operationId: Subscribe
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/billing.SubscribeRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/billing.SubscriptionResponse"
          description: OK
...
  /billing/webhook:
    post:
      operationId: Webhook
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/billing.WebhookRequest"
        required: true
      responses:
        "200":
          description: OK
...
  /handle/{handle}:
    get:
      operationId: GetPublicBadgeByHandle
      parameters:
      - in: path
        name: handle
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/report.PublicBadge"
          description: OK
...
  /influencers/{id}/handles:
    post:
      operationId: AddHandle
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/influencer.AddHandleRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/influencer.HandleResponse"
          description: OK
  /influencers/{id}/handles/{handleID}:
    delete:
      operationId: DeleteHandle
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
      - in: path
        name: handleID
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
...
  /waitlist:
    post:
      operationId: Capture
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/waitlist.CaptureRequest"
        required: true
      responses:
        "200":
          description: OK
  /whitelabel:
    get:
      operationId: GetWhitelabel
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/whitelabel.WhitelabelResponse"
          description: OK
    put:
      operationId: UpdateWhitelabel
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/whitelabel.UpdateWhitelabelRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/whitelabel.WhitelabelResponse"
          description: OK
servers:
- url: /v1